requirements.txt
nba_api
pandas
numpy
requests
python-dotenv
scipy
cloudscraper
